toolkit/mozapps/downloads/tests/chrome/test_bug_429247.xul

<?xml version="1.0"?>
<!--
/* ***** BEGIN LICENSE BLOCK *****
 * Version: MPL 1.1/GPL 2.0/LGPL 2.1
 *
 * The contents of this file are subject to the Mozilla Public License Version
 * 1.1 (the "License"); you may not use this file except in compliance with
 * the License. You may obtain a copy of the License at
 * http://www.mozilla.org/MPL/
 *
 * Software distributed under the License is distributed on an "AS IS" basis,
 * WITHOUT WARRANTY OF ANY KIND, either express or implied. See the License
 * for the specific language governing rights and limitations under the
 * License.
 *
 * The Original Code is Download Manager UI Test Code.
 *
 * The Initial Developer of the Original Code is
 * Edward Lee <edward.lee@engineering.uiuc.edu>.
 * Portions created by the Initial Developer are Copyright (C) 2008
 * the Initial Developer. All Rights Reserved.
 *
 * Contributor(s):
 *  Nochum Sossonko <nsossonko@hotmail.com>
 *
 * Alternatively, the contents of this file may be used under the terms of
 * either the GNU General Public License Version 2 or later (the "GPL"), or
 * the GNU Lesser General Public License Version 2.1 or later (the "LGPL"),
 * in which case the provisions of the GPL or the LGPL are applicable instead
 * of those above. If you wish to allow use of your version of this file only
 * under the terms of either the GPL or the LGPL, and not to allow others to
 * use your version of this file under the terms of the MPL, indicate your
 * decision by deleting the provisions above and replace them with the notice
 * and other provisions required by the GPL or the LGPL. If you do not delete
 * the provisions above, a recipient may use your version of this file under
 * the terms of any one of the MPL, the GPL or the LGPL.
 *
 * ***** END LICENSE BLOCK ***** */

/**
 * Test bug 429247 to make sure clicking on whitespace in the Download
 * Manager doesn't run the default action on the selected download.
 */
-->

<window title="Download Manager Test"
        xmlns="http://www.mozilla.org/keymaster/gatekeeper/there.is.only.xul"
        onload="test();">

  <script type="application/javascript"
          src="chrome://mochikit/content/tests/SimpleTest/SimpleTest.js"/>
  <script type="application/javascript"
          src="chrome://mochikit/content/tests/SimpleTest/EventUtils.js"/>
  <script type="application/javascript"
          src="utils.js"/>

  <script type="application/javascript">
  <![CDATA[

function test()
{
  var dmui = getDMUI();
  if (!dmui) {
    todo(false, "skip test for toolkit download manager UI");
    return;
  }

  let ios = Cc["@mozilla.org/network/io-service;1"].getService(Ci.nsIIOService);
  let dmFile = Cc["@mozilla.org/file/directory_service;1"].
               getService(Ci.nsIProperties).get("TmpD", Ci.nsIFile);
  dmFile.append("dm-ui-test.file");
  dmFile.createUnique(Ci.nsIFile.NORMAL_FILE_TYPE, 0666);
  let gTestPath = ios.newFileURI(dmFile).spec;

  // Data to populate the download manager with
  const DownloadData = {
    name: "381603.patch",
    source: "https://bugzilla.mozilla.org/attachment.cgi?id=266520",
    target: gTestPath,
    startTime: 1180493839859230,
    endTime: 1180493839859239,
    state: Ci.nsIDownloadManager.DOWNLOAD_FINISHED,
    currBytes: 0, maxBytes: -1, preferredAction: 0, autoResume: 0
  }

  let dm = Cc["@mozilla.org/download-manager;1"].
           getService(Ci.nsIDownloadManager);
  let db = dm.DBConnection;

  // Empty any old downloads
  db.executeSimpleSQL("DELETE FROM moz_downloads");

  let stmt = db.createStatement(
    "INSERT INTO moz_downloads (name, source, target, startTime, endTime, " +
      "state, currBytes, maxBytes, preferredAction, autoResume) " +
    "VALUES (:name, :source, :target, :startTime, :endTime, :state, " +
      ":currBytes, :maxBytes, :preferredAction, :autoResume)");
  for (let prop in DownloadData)
    stmt.params[prop] = DownloadData[prop];

  stmt.execute();
  stmt.finalize();

  // Close the UI if necessary
  let wm = Cc["@mozilla.org/appshell/window-mediator;1"].
           getService(Ci.nsIWindowMediator);
  let win = wm.getMostRecentWindow("Download:Manager");
  if (win) win.close();

  let obs = Cc["@mozilla.org/observer-service;1"].
            getService(Ci.nsIObserverService);
  const DLMGR_UI_DONE = "download-manager-ui-done";

  let testObs = {
    observe: function(aSubject, aTopic, aData) {
      if (aTopic != DLMGR_UI_DONE)
        return;

      let win = aSubject.QueryInterface(Ci.nsIDOMWindow);
      win.focus();
      let $ = function(aId) win.document.getElementById(aId);
      let downloadView = $("downloadView");

      // Default test/check for invocations
      let invokeCount = 0;
      let counter = function() invokeCount++;

      // Array of tests that consist of the download manager
      // function to temporarily replace, method to use in its place, value to
      // use when checking correctness
      let defaultForSelectedTest = ["doDefaultForSelected", counter, counter];

      // All the expected results
      let allExpected = [0, "The default action was not performed"];

      // Select the first download
      downloadView.selectedIndex = 0;

      let [func, test, value] = defaultForSelectedTest;
      // Make a copy of the original function and replace it with a test
      let copy = win[func];
      win[func] = test;

      // Get the offsetY of the whitespace under the download
      let downloadItem = downloadView.getItemAtIndex(0);
      let boxOffset = downloadItem.boxObject.height + 1;
      // The last download in the download manager + 1 is whitespace
      synthesizeMouse(downloadItem, 0, boxOffset, {clickCount: 2}, win);

      // Make sure the value is as expected
      let [correct, message] = allExpected;
      ok(value() == correct, message);

      // Restore original values
      invokeCount = 0;
      win[func] = copy;

      // We're done!
      win.close();
      obs.removeObserver(testObs, DLMGR_UI_DONE);
      SimpleTest.finish();
    }
  };
  obs.addObserver(testObs, DLMGR_UI_DONE, false);

  // Show the Download Manager UI
  dmui.show();

  SimpleTest.waitForExplicitFinish();
}

  ]]>
  </script>

  <body xmlns="http://www.w3.org/1999/xhtml">
    <p id="display"></p>
    <div id="content" style="display:none;"></div>
    <pre id="test"></pre>
  </body>
</window>
